* Fixed some bugs, memory leaks and crashes in the registration code, RSA code, net code * Wrote test code for registration
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzauthd@18092 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ bzauthd/RSA.h
@@ -31,8 +31,8 @@ public:
   friend class RSAManager;
   RSAKey();
   virtual ~RSAKey();
-  bool setValues(uint8 *n, size_t n_len, uint16 e);
-  bool getValues(uint8 *&n, size_t &n_len, uint16 &e);
+  bool setValues(uint8 *n, size_t n_len, uint32 e);
+  bool getValues(uint8 *&n, size_t &n_len, uint32 &e);
   virtual RSAKeyType getType() = 0;
 protected:
   uint8 *_getValueN(size_t *n_len);
